+++ SampleShopApp/Networking/Parsers/TMDBMoviesListParser.swift
@@ -19,7 +19,7 @@ class TMDBMoviesListParser: TMDBBaseParser {
         
         let resultObject = TMDBMovieListResultItem()
         resultObject.currentPage = rawDict["page"] as? Int
-        resultObject.totalPages = rawDict["total_pages"] as? Int64
+        resultObject.totalPages = rawDict["total_pages"] as? Int
         
         if let rawResults = rawDict["results"] as? Array<Dictionary<String, Any>> {
